+++ src/core/shaderlib/project64/project64.js
@@ -20,7 +20,7 @@
 
 import {fp64} from 'luma.gl';
 import project from '../project/project';
-import {getFP64ViewportUniforms} from './viewport-uniforms-64';
+import {fp64ify, fp64ifyMatrix4} from '../../utils/fp64';
 
 import project64Shader from './project64.glsl';
 
@@ -41,5 +41,19 @@ export default {
 const DEFAULT_MODULE_OPTIONS = {};
 
 function getUniforms(opts = DEFAULT_MODULE_OPTIONS, context = {}) {
-  return getFP64ViewportUniforms(context);
+  // These two uniforms are generated by the fp32 project module
+  const {project_uViewProjectionMatrix, project_uScale} = context;
+
+  if (project_uViewProjectionMatrix && project_uScale) {
+    // We only need to update fp64 uniforms if fp32 projection is being updated
+    const glViewProjectionMatrixFP64 = fp64ifyMatrix4(project_uViewProjectionMatrix);
+    const scaleFP64 = fp64ify(project_uScale);
+
+    return {
+      project_uViewProjectionMatrixFP64: glViewProjectionMatrixFP64,
+      project64_uViewProjectionMatrix: glViewProjectionMatrixFP64,
+      project64_uScale: scaleFP64
+    };
+  }
+  return {};
 }
